NEW CAMPAIGN
It's here. A new campaign map is here. You arrive at the Bay of Discovera, on a island called Tropica and have to conquer all of the arenas, before encountering the world champion, and defeating them, to become the world champion. We promise a big cave for every country in Tropica, (aside from Manta, the tutorial area) with unique materials and fun events! Anyways, it's still a W.I.P, but still, cool!